+++ src/GafferModule/CompoundNumericPlugBinding.cpp
@@ -58,7 +58,7 @@ namespace
 {
 
 template<typename T>
-std::string serialisationRepr( const T *plug, const Serialisation *serialisation = nullptr )
+std::string serialisationRepr( const T *plug, Serialisation *serialisation = nullptr )
 {
 	std::string extraArgs = "";
 
@@ -87,7 +87,7 @@ class CompoundNumericPlugSerialiser : public ValuePlugSerialiser
 
 	public :
 
-		std::string constructor( const Gaffer::GraphComponent *graphComponent, const Serialisation &serialisation ) const override
+		std::string constructor( const Gaffer::GraphComponent *graphComponent, Serialisation &serialisation ) const override
 		{
 			return serialisationRepr( static_cast<const T *>( graphComponent ), &serialisation );
 		}