+++ src/GafferModule/SerialisationBinding.cpp
@@ -90,7 +90,7 @@ IECore::ObjectPtr objectFromBase64Wrapper( const std::string &base64String )
 void GafferModule::bindSerialisation()
 {
 
-	scope s = boost::python::class_<Serialisation>( "Serialisation", no_init )
+	scope s = boost::python::class_<Serialisation, boost::noncopyable>( "Serialisation", no_init )
 		.def(
 			init<const Gaffer::GraphComponent *, const std::string &, const Gaffer::Set *>
 			(
@@ -104,10 +104,11 @@ void GafferModule::bindSerialisation()
 		.def( "parent", &parent )
 		.def( "identifier", &Serialisation::identifier )
 		.def( "childIdentifier", &childIdentifier )
+		.def( "addModule", &Serialisation::addModule )
 		.def( "result", &Serialisation::result )
-		.def( "modulePath", (std::string (*)( object & ))&Serialisation::modulePath )
+		.def( "modulePath", (std::string (*)( const object & ))&Serialisation::modulePath )
 		.staticmethod( "modulePath" )
-		.def( "classPath", (std::string (*)( object & ))&Serialisation::classPath )
+		.def( "classPath", (std::string (*)( const object & ))&Serialisation::classPath )
 		.staticmethod( "classPath" )
 		.def( "objectToBase64", &objectToBase64Wrapper )
 		.staticmethod( "objectToBase64" )